requirements/requirements-testing.txt
coverage
django-nose
django-dynamic-fixture
psycopg2
flake8
